# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.BranchData:
    properties:
      address:
        example: "100"
        type: string
      branchTxnTblName:
        example: "12"
        type: string
      createdAt:
        example: "2006-01-02 23:59:59"
        type: string
      createdBy:
        example: shiva
        type: string
      district:
        example: "12"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: branch_2022
        type: string
      pinCode:
        example: "12"
        type: string
      place:
        example: "1000"
        type: string
      remarks:
        example: test
        type: string
      state:
        example: "12"
        type: string
      status:
        example: ACTIVE
        type: string
      taluka:
        example: "12"
        type: string
      updatedAt:
        example: "2006-01-02 23:59:59"
        type: string
      updatedBy:
        example: shiva
        type: string
    type: object
  model.BranchsResponse:
    properties:
      code:
        example: 400
        type: integer
      data:
        items:
          $ref: '#/definitions/model.BranchData'
        type: array
      message:
        example: status bad request
        type: string
    type: object
host: localhost:3081
info:
  contact:
    email: shivukumar076@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: Loan Sanction Wirte server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Loan Sanction API
  version: "1.0"
paths:
  /getBranchByBranchId/{branchId}:
    get:
      consumes:
      - application/json
      description: Branch
      parameters:
      - description: Branch Id
        in: path
        name: branchId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BranchsResponse'
      summary: Branch
      tags:
      - Branch
  /getBranchByBranchName/{branchName}:
    get:
      consumes:
      - application/json
      description: Branch
      parameters:
      - description: Branch Name
        in: path
        name: branchName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BranchsResponse'
      summary: Branch
      tags:
      - Branch
  /getBranchs/{page}/{size}:
    get:
      consumes:
      - application/json
      description: Branch
      parameters:
      - description: Page Number
        in: path
        name: page
        required: true
        type: string
      - description: Per Page Records
        in: path
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.BranchsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.BranchsResponse'
      summary: Branch
      tags:
      - Branch
  /ping:
    get:
      description: Ping
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: error
          schema:
            type: string
      summary: Ping
      tags:
      - Ping
swagger: "2.0"