# src/main/go/model/BranchResponse.go
/*
 * Loan
 *
 * Fd Pack read docs
 *
 * API version: 0.0.1
 * Contact: shivakumar@supernet-india.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type BranchResponse struct {
	Code    int        `json:"code" example:"400"`
	Message string     `json:"message" example:"status bad request"`
	Data    BranchData `json:"data"`
}

type BranchsResponse struct {
	Code    int          `json:"code" example:"400"`
	Message string       `json:"message" example:"status bad request"`
	Data    []BranchData `json:"data"`
}

type BranchData struct {
	Id               int64  `json:"id,omitempty" example:"1"`
	Name             string `json:"name,omitempty" example:"branch_2022"`
	Address          string `json:"address,omitempty" example:"100"`
	Place            string `json:"place,omitempty" example:"1000"`
	Taluka           string `json:"taluka,omitempty" example:"12"`
	District         string `json:"district,omitempty" example:"12"`
	State            string `json:"state,omitempty" example:"12"`
	PinCode          string `json:"pinCode,omitempty" example:"12"`
	BranchTxnTblName string `json:"branchTxnTblName,omitempty" example:"12"`
	Status           string `json:"status,omitempty" example:"ACTIVE"`
	CreatedAt        string `json:"createdAt,omitempty" example:"2006-01-02 23:59:59"`
	UpdatedAt        string `json:"updatedAt,omitempty" example:"2006-01-02 23:59:59"`
	CreatedBy        string `json:"createdBy,omitempty" example:"shiva"`
	UpdatedBy        string `json:"updatedBy,omitempty" example:"shiva"`
	Remarks          string `json:"remarks,omitempty" example:"test"`
}